refactor(kbar): replace self-contained SSE impl with Crow headless script tag
Instead of owning the SSE streaming logic in useCrowChat.ts, load the
Crow widget via a Next.js Script tag (data-headless="true") following
the same pattern as FreshChat. CrowFallback now calls window.crow('ask')
with callbacks, delegating all networking to the SDK.

- Add CrowScript.tsx (23 lines) — opt-in via NEXT_PUBLIC_CROW_API_URL
- Rewrite CrowFallback to use window.crow('ask') with local state
- Delete useCrowChat.ts (220 lines)
